Replace Map mutations with dict mutations for purity
Summary:
D31595807 (https://github.com/facebook/fbthrift/commit/b92d34e56406a1b1e56cb6445bd8fcc379772e92) replaced a call to `Map::fromItems()` with a `foreach` loop that mutates a brand-new map. This is a mutation, so the function is not considered pure.

Of course, `Map::fromItems()` also creates a map and mutates it by inserting stuff... but this happens in C++ code, so enforcement doesn't care.

This inconsistency reinforces my belief that purity is not ready for WWW - mutating a brand-new map, without making it observable to anything else outside the function, should not be something we allow in C++ but not in Hack.

Instead of mutating a Map, we can mutate a dict and convert to a Map later. Dict mutations are considered pure, as the dict will be copied if it is held by another function.

Differential Revision: D31809452

fbshipit-source-id: 412c3c8bfe593ace692037b6ee9acd98fa2228c1