Ensure executor lifetime continuity in PooledRequestChannel
Summary: We currently drop the shared_ptr to the IOExecutor between getting an EventBase from it and adding work to that EventBase, which leaves open a race condition where the threadpool is destroyed before (or concurrently with) the add call. Returning a KeepAlive provides a continuous guarantee, and we were usually constructing this keepalive later anyway.

Differential Revision: D31908280

fbshipit-source-id: 58116839e6f892672b784338b18abf8937cb2b18