Cap the number of credits to chunkBufferSize
Summary:
Today with memory buffer size based flow control we compute the number of
credits based on a historial estimate of maxPayloadSize. This works well
for use cases that have a uniform traffic pattern but doesn't work well
when the traffic bursty. For example, we have long periods of small
payloads (heart-beats for example) followed by a spike of really large
payloads. In such a scenario, our computation of outstanding credits
might overwhelm the receiver. This diff caps the number of credits to
the `initialChunks` value provided as part of RpcOptions::setMemoryBufferSize

Reviewed By: iahs

Differential Revision: D31908414

fbshipit-source-id: 783f997fe377a78a58f92c417321397e756f8edd