Create cpp library to wrap around py3lite server implementation
Summary:
This creates a wrapper meant to interface with the current thrift cpp server implementation without needing cpp2 codegen.

This is done by accepting a map of the python methods to be called and generating the normal cpp2 function metadata on the fly from this map. In addition, since serialization/deserialization here is being moved into the python code (because you need code gen to be able to handle that), the expected return type on the handler callback is IOBuf, since the result will be returned serialized.

Oneway functions are also special cased here (even though this does bother me) because they have different lifecycles where they have a promise that accepts folly::Unit and don't send a reply back to the client.

The cython files here use the current Thrift Server implementation, but replace the AsyncProcessor used there with this custom one I wrote.

Reviewed By: praihan

Differential Revision: D31325558

fbshipit-source-id: 1e9de411401e491663c90e7a7c076483e7727a36