bugfix: lazy cpp.ref optional field might not be serialized
Summary:
Currently if field is cpp.ref and optional, we will check whether unique_ptr == nullptr to decide whether to serialize this field. (or consider this field as empty)

However, there is a bug here -- if field is lazy, we will delay the deserialization, thus the unique_ptr will be nullptr despite lazy IOBuf is not empty.

The fix is checking whether this field has undeserialized data, if so, always serialize it.

NOTE: This doesn't affect non cpp.ref field, since in that case __isset is always updated regardless whether actual data is lazily deserialized.

NOTE: Currently we don't have such use case in production per fbgs.

Reviewed By: iahs

Differential Revision: D32038083

fbshipit-source-id: caa7837a8b7a9f085563309a8821565053c5fd23