Only instantiate getName function if used.
Summary: Previously this function was on the type tag, which causes the compiler to instantiate it whenever the type tag is used. This causes problems when reflection, used by this helper function, is not included. Moving this to a ~free function solves the problem by only requiring reflection if the function is actually used.

Reviewed By: Mizuchi

Differential Revision: D31458822

fbshipit-source-id: e4e3ffa1769c74eec580927cf14269613c9881df