Fix flaky test
Summary:
`ScopedServerInterfaceThread::newClient()` allocates clients backed by a
`PooledRequestChannel` and folly's global IO executor. This means that
requests may go through different channels, even when a single client is used,
so the order in which requests are sent is not guaranteed. This test assumed a
specific request order, causing it to fail every 100 runs or so.

Reviewed By: AkramaMirza

Differential Revision: D31917137

fbshipit-source-id: 9d241b92c1b846b897650fe14d4086d70966995d