Create mustache templates for py3lite servers
Summary:
Adds template files for py3lite servers, which is done by implementing a `_fbthrift__handler__{name}` function for every normal function on the service, which gets passed as a part of the function table to the cpp server.

These handler functions deserialize the args, call the normal server function, then serialize the result again and return it back to the cython layer.

Reviewed By: nanshu

Differential Revision: D31329118

fbshipit-source-id: 8fd4f1e1323b7afcd82df14c341445ba6d9ae851