Refactor some types and SFINAE usage
Summary:
**This stack:**
The goal of this stack is to introduce a framework/utility to produce hashes in a deterministic way.
The framework will allow users to control the hashing mechanism and will visit thrift structs in a consistent and stable manner.
Users will implement a small class that will dictate how to produce partial hashes for thrift primitive types.

**This diff:**
Fix some type usages: use typed ints instead of chars for non-char data types, use the right type casts in tests.
Fix SFINAE to be a template non-typename argument and not a return type.
Remove some raw array usages.

Reviewed By: yfeldblum

Differential Revision: D32003770

fbshipit-source-id: cacaf36c2ffab486aabe46549158421e9a1a1769