Compiler changes for hack.ExperimentalAdapter
Summary:
Add the `hack.ExperimentalAdapter` (naming to be confirmed) annotation that applies a hack wrapper to fields. It has similar functionality as `hack.adapter` but it wraps the field instead of changing to a fixed hack type.

For example:

```
struct MyStruct {
  edward-io.ExperimentalAdapter{name = "\MyFieldAdapter", adapted_generic_type = "\MyFieldWrapper"}
  1: i64 wrapped_field;
}
```

If `adapted_generic_type` is not provided then it will behave similar to existing `hack.adapter` annotation and use the old APIs

Reviewed By: prasad223

Differential Revision: D31045292

fbshipit-source-id: a807d7e958d7bf11684bf33783a8adf5a784818e