Replace std::mutex with folly::SharedMutex in lazy deserialization
Summary:
`std::mutex` has a 40 byte impact on class footprint, while `folly::SharedMutex` is only 4 bytes (and can be laid out to not introduce any padding).

Furthermore, we should not mutually exclude `const` methods, in particular copy constructor and serialization, so we can use a read lock for those.

In terms of performance, in the uncontended case `SharedMutex` is actually faster because the fast path can be inlined. In the contended case, the critical section is always large enough (involving allocations) that the waiters would go to sleep anyway, so any mutex will have about the same performance.

Reviewed By: Mizuchi, philippv, luciang, vitaut

Differential Revision: D32429347

fbshipit-source-id: db10b4f2b2ebb801b58dd9d26901d56b2636aa97