Add option to use an IOBuf ref for lazy fields
Summary:
Overhead of a lazy field is currently 57 bytes (`IOBuf` + flag) plus padding, which is unacceptable in classes with several fields out of which small subsets are set for each object. This impacts also the cost of all the constructors and the destructor.

Add an option in the `Lazy` annotation to swap the `IOBuf` with a `std::unique_ptr`. The trade off is an extra allocation when deserializing, but that can be largely offset by removing the overhead, depending on the nature of the data.

Reviewed By: luciang

Differential Revision: D32430246

fbshipit-source-id: 2ac91d5dad8ad4c3bd1a41200cf12ac1ff56cbc6