Fix onFirstResponse flow if it's an exception
Summary:
According to the contract defined in StreamCallbacks.h, `onFirstResponse` should only return false if it calls a terminating method on the server callback inline:
https://www.internalfb.com/code/fbsource/[633744e17d9637563967b818dc5ac98c177cd7ad]/fbcode/thrift/lib/cpp2/async/StreamCallbacks.h?lines=206-208

For example RTI does this:
https://www.internalfb.com/code/fbsource/[633744e17d96]/fbcode/realtime/bladerunner/client/transport/thrift/StreamHandle.cpp?lines=25-37

However, ClientStreamBridge and ClientSinkBridge do not conform to this since they return false if the first payload is an exception without calling a terminating method on the server callback. This diff fixes it by always returning true.

With this change, RocketClient will first call `onFirstResponse` followed by `onStreamComplete`/`onFinalResponse` if `onFirstResponse` returned `true`.

Reviewed By: iahs

Differential Revision: D32438498

fbshipit-source-id: 627e7fffdfb18a6bc5417f1121e584078aafa377