bugfix: compare value of cpp.ref field in union instead of address
Summary:
There is bug in union's `operator<` -- if field is `cpp.ref`, we compare the memory address instead of value.

This diff fixed the problem by dereferencing `cpp.ref` field in `operator<`.

It's unlikely existing code will rely on this bug since the result is usually random when we compare memory address.

This diff also added general unit-test `CppRefUnionSetterGetter`, since we don't have it for `cpp.ref` field today.

Reviewed By: iahs

Differential Revision: D32241268

fbshipit-source-id: 6431ad82fbe543f45c9ddf7578d1f105b87fe06e