Add void_t and is_thrift_type_tag
Summary:
`void_t` is for use as a 'default' type, and for function return types.
`is_thrift_type_tag` is for use in static asserts and std::enable_if.

also rename is_concrete_type to is_concrete (the _type was left over from the previous version of the lib)
and move it directly into ThriftType.h, as it simplifies the implementation of is_thrift_type_tag, is simple to compute, and is a useful thing to put in a static assert.

Reviewed By: Mizuchi

Differential Revision: D32577959

fbshipit-source-id: bc371d352a2629c71866de32283ca437bb264f66