Add a minRtt limit to the AdaptiveConcurrency controller
Summary:
AdaptiveConcurrency controller computes a new target RTT
latency by clamping down the server's concurrency to the minimum
value and measuring request processing times during this period. The
p95 of measured value is selected as the new target (with some buffer).
However, this might not be realistic in certain cases where the service
may not have strict processing latency requirements and may prioritize
throughput over latency.

To address this drawback, we're adding a configuration knob to set
a lower limit on the target RTT latency. This can be configuraable via
configurator like the other existing knobs. The logic is fairly simple -
if the value is non-zero then we take the `max(minTargetRTT, computedRtt)`
as the new ideal RTT latency target.

Also includes a small potential divide by zero fix in the `raw_gradient` calculation.

Reviewed By: praihan

Differential Revision: D32609249

fbshipit-source-id: cfc30debe6f0ee8b08b1c0a439af49bb46cfd8bf