Support type tag-based function overloading
Summary: When there are a large number of specializations, this method is prefered over a std::enable_if based approach, which requires the compiler to examine each specialization. On the other hand, since compilers are highly optimized for overload resolution, an overload-based approach is almost always a ~constant time lookup.

Reviewed By: thedavekwon

Differential Revision: D32714946

fbshipit-source-id: bd31afc70bed005a830bf164188e1d312dcb065a