Extract implementation of properties into cdef function so it can be directly consumed by Cython code
Summary:
Currently Thrift Cython objects do expose thrift properies as Python `property` - this is a perfectly valid way of doing this however when this object is used from Cython code it adds an extra cost for the attribute lookup.

As a proposal we now put implementation of the property getter into `cdef` function with `_impl` siffix in Cython which is called from the actual property code. This allows Cython consumers to access `cdef` instead of property

Reviewed By: nanshu

Differential Revision: D32536424

fbshipit-source-id: 1bad024e4c4f9f6fa22e567a80161ba4a8a57251