Make the ideal target RTT samping percentile configurable
Summary:
Adaptive Concurrency controller currently has the following
configurable parameters:
- minConcurrency
- targetRttFactor (to account for bursts)
- targetRttFixed
- minTargetRtt

One additional parameter that should potentially be configurable is the percentile
value used to select the target RTT latency from the sample set of requests
captured during the recalibration window. This is currrently hardcoded to p95
which is a reasonable default but we'd like to make it configurable so that
we can experiment more easily with the target latency computation.

The target percentile will be limited to the range (0.0, 1.0) exclusive.

Reviewed By: sethdelliott

Differential Revision: D32735066

fbshipit-source-id: 48a1b7ea0b199c386d83f4c78641ccc77b26288a