Correctly handle StopTLS error scenarios
Summary:
The existing implementation of StopTLS is not completely correct with respect
to `dropConnection()`. This will result in a shutdown crash if any Thrift
server is stopping _while there exists any pending StopTLS connections_.

This diff handles the `dropConnection()` path by ensuring that a read callback
is attached while a stoptls transaction is in progress. This allows us to
receive transport level errors while we wait for the corresponding close_notify
to arrive.

Reviewed By: knekritz

Differential Revision: D32338110

fbshipit-source-id: 448143186b165a429a691fc2a46d43b65532ed27