Add an option to TableBasedSerializer to use IOBuf from ThriftValue to read/write data
Summary:
Currently the way TableBasedSerializer works with IOBufs is that it just static casts what's stored in the data field into an IOBuf and tries to use it.

This doesn't work for py3lite because what would be stored there is a python object and not a c IOBuf, so the `.get` api has to be used, but this runs into a problem that there's no part of this API that returns an IOBuf, so to avoid a breaking change for other uses of TableBasedSerializer I added another StringFieldType category to use the result of getValue.

Unfortunately this means that there are now 3 separate IOBuf related enums, but that seems necessary for the various implementations.

Reviewed By: vitaut

Differential Revision: D32403863

fbshipit-source-id: c55a134c123627865dff381adc11fa4e1ab10f3e