Add IOBuf support to py3lite
Summary:
Adding a new type info for IOBuf, and having it be converted to a byte array when serializing (not sure if there's a better way of leaving it as a cIOBuf but this works).

Adding a new type annotation for IOBuf which is specific to python (py3.iobuf) so we don't have to rely on cpp annotations, but still leaving in recognition for the cpp.type annotation for the sake of compatibility with py3.

Reviewed By: nanshu

Differential Revision: D32266764

fbshipit-source-id: ca3923120a24233dfebd39704cc93c6701091ce9