Change public `__clear` to private `__fbthrift_clear` member function and access via `struct_private_access`
Summary:
`__clear` uses a reserved name which is a UB conflicting standard.
`__clear` has a use case of clearing contents without avoiding deallocating memory.

Currently `apache::thrift::clear` is implemented as a member invoker for `__clear` member method. Rename `__clear` to `__fbthrift_clear` and make it as a private function.

Reviewed By: iahs

Differential Revision: D32961403

fbshipit-source-id: 5304f36c4cf4632dfb2b47152eafbba640548690