Fix crash in py3 serializer when using deserialize_from_header
Summary:
The function `THeader.removeHeader` can throw an exception but the no exceptions are declared in the relevant pxd file so it crashes the program.

Also while writing a unit test with bad data, this function returned a null IOBuf, which caused a segfault later on, so checking for null IOBuf as well.

Reviewed By: yfeldblum, nanshu

Differential Revision: D32968228

fbshipit-source-id: 14474e657e2f0319946f898d61e6e28bbab3718e