Use default coeffects for field adapter methods
Summary:
Facebook :

Field adapter test in D31457295 is failing because of coeffects mismatch between hhvm and generated setter methods for field adapter.

In this diff, I am changing hhvm methods to use `defaults` instead of pure and similarly, using `defaults` for all the field adapter methods.

I am not using `write_props` since it doesn't work well with `policied_shallow`. As seen in diff D32451070, `defaults` works with `policied_shallow` and most likely we will have to change it to `defaults` later anyways. To avoid additional wait for another hhvm release, I am changing this to `defaults` right away. If I find a better way to customize it, I'll do it in separate diff. This will atleast unblock D31457295 and also make it easier to test with `policied_shallow`.

Reviewed By: prasad223

Differential Revision: D33110250

fbshipit-source-id: 74abd3db1b159ddce3f317e0feea82d350c10f71