upgrade to version 3.19.1 in both fbsource and platform009
Summary:
allow-large-files

This diff is upgrading both `fbsource/third-party/protobuf` and `platform009`
protobuf to version `3.19.1` which actually unblocks building `Eden` on
`platform010`.

The `fbsource/third-party/protobuf` is intertwined with the `platform009` /
`platform010` protobuf due to which we need to upgrade all 3 together.

This diff also upgrades `grpc` to version `1.42.0` on `platform009`.

Reviewed By: digdoug

Differential Revision: D33028221

fbshipit-source-id: efbbe009b998ed23927ba8b7f0cb735ce9b641e1