Support matcher for unions with `IsThriftUnionWith`
Summary:
There is currently no matcher for unions.
This diff introduces a matcher equivalent to `IsVariantWith`, which is provided by gtest.

The matcher matches the variant if the active member is the same as the expected one, and the value of the member matches the inner matcher provided to the `IsThriftUnionWith`.
An unset union never matches.

See below the error being printed when the matching fails.

Given

```
auto r = Result();
r.set_error("error");
EXPECT_THAT(r, IsThriftUnionWith<field::success>(42));
```
it prints
```
Value of: r
Expected: is an union with `success` active, which is equal to 42
  Actual: 40-byte object <30-10 11-BA B3-7F 00-00 05-00 00-00 00-00 00-00 65-72 72-6F 72-00 00-00 00-00 00-00 00-00 00-00 02-00 00-00 00-00 00-00> (of type apache::thrift::test::Result), whose active member `error` is "error"
```

When the `_fatal_union.h` is not included, the compilation fails with a useful error message:

```
thrift/lib/cpp2/test/Matcher-inl.h:148:5: error: static_assert failed due to requirement 'SupportsReflection<apache::thrift::test::Result>::value' "Include the _fatal_union.h header for the Thrift file defining this union"
```

Reviewed By: iahs

Differential Revision: D33090727

fbshipit-source-id: 891ea44b9804b14138c32e6970e04e9543eae934