remove java.swift.disable_reactive
Summary:
The motivation of introducing this annotation is no longer there, since no one is using this for now, we can remove it to make things a bit simpler.

In the future, the reactive client will be the default generated client (blocking and async clients are wrapper around this client), we should not have a switch to turn it off now.

Reviewed By: prakashgayasen

Differential Revision: D33269626

fbshipit-source-id: a9bc3c7afaeade4ec2bd01c7b36476b434fb48f5