Put in a test as an example for how RequestContext in thrift py3 streaming works
Summary:
Unfortunately this awkward pattern is required to use RequestContext in server streaming, because the context only exists during the lifetime of the original request, and in python if you have any yields in a function the whole thing becomes a generator (unlike in c++ where you can first execute some code and then return a generator).

This means that to acccess the request context before it's destroyed, you have to access it first in one method then call a generator method in a pretty awkward pattern (I can't come up with any way to avoid this unfortunately).

Reviewed By: nanshu

Differential Revision: D33284370

fbshipit-source-id: 1103c95e8e8d81093ffe422eea3a6ef19c5aba83