Drop logic that depends on RequestClientCallback::onRequestSent being called for twoway RPCs
Summary:
Rocket returns NOT_OPEN errors for requests that were not fully written to the socket. But it only calles onRequestSent when the full batch is written into the socket. So it makes relying on both onRequestSent timing or presence (e.g. to distinguish between send vs recv errors) incorrect.
This change eliminates such dependency an only relies on the error code.

Reviewed By: iahs

Differential Revision: D33194179

fbshipit-source-id: be1325cd1b53c29634268e4d2fea49f8a898ed62