Stop calling onRequestSent for two-way requests
Summary:
onRequestSent doesn't have clear semantics in terms of both timing and which errors it doesn't get called for (e.g. does not getting onRequestSent guarantee that request was never dispatched to any server?), so it should be dropped.
Previous diffs already ensure that no code depends on onRequestSent calls.

Reviewed By: iahs

Differential Revision: D33198872

fbshipit-source-id: 33e963b3491f9731f51e6766a2bf6a489c3561bc