Remove None from possible return types from py3lite struct __iter__
Summary:
Currently, this uses `{{ > types/field_value_pep484_union_type}}`, which includes a `None` type, making this union effectively optional, which means that when these fields are being iterated through every individual field is marked optional even though none of them are.

None is still left in as a base case if the struct has no fields as `Union[]` is invalid syntax.

Reviewed By: nanshu

Differential Revision: D33107989

fbshipit-source-id: cacae515c7218a7a49d31c218239cb948481167e