Hook up op::hash to the type-parameterized ThriftOp tests.
Summary:
And validate that all 'interesting' key values have unique hashes for the default hasher using the ValueGenerator.

Note that the number of type tags being tested is relatively limited. A follow up diff should be created that also uses the ValueGenerator, but with a more comprehensive list of Thrift type tags, to generate 'golden' hashes for a wide range of types and value, to verify the deterministic hashing logic is stable as we evolve the code. An example of such a test can be found in thrift/test/testset:generator_test or thrift/compiler/test:compiler_test

Reviewed By: thedavekwon

Differential Revision: D33251320

fbshipit-source-id: 5d26d4cf275e47942b0b6a2b3abee260483953ba