Migrate py clients from un/framed to header transports
Summary:
Header transport is supported by all known thrift server implementations. Un/framed transports do not provide the minimum level of telemetry required and are super-deprecated (as opposed to header which is just regular deprecated).
Update channel to replace these transports with header. Unlike with rocket upgrade, there's no negotiation because we don't know of any thrift servers to add it to that don't already support the header transport.

Reviewed By: joshkirstein

Differential Revision: D33191727

fbshipit-source-id: e22acd04ba5913c54ca62fcef464abc87168bcdd