Fix contract violation in sink final response path
Summary: There are two ways to get to the final response section modified in this diff: the generator completing or the server sending a final response. In the latter case cancellation is meaningless because there's nothing to wait for, but in the former case we are waiting for the server to send a final response and that wait can be cancelled. We would like to tell the server we no longer care about the final response, but the sink contract does not allow it since we already called onSinkComplete. Remove the forbidden error message.

Differential Revision: D33303175

fbshipit-source-id: 2365c3f795f0135788e71b818b815cbf79fac74a