Remove _VALUES_TO_NAMES _NAMES_TO_VALUES in thrift enum code for Android
Summary: _VALUES_TO_NAMES and _NAMES_TO_VALUES are initialized at library load time for each type.  In Assistant Client Runtime, we use tons of thrift types, so each time we load our library we have ~100 initializers to run.  This causes a lot of the library to be loaded which causes a lot of page faults and slows down startup.  These are not used anywhere in xplat or fbandroid, so we can just compile them out.  This improves startup time for Milan Speech Service for example.

Reviewed By: Mizuchi

Differential Revision: D33489019

fbshipit-source-id: 5ce627f57f25c75f4f4d51016b1f9fbabee07bd0