Add AnyRef
Summary:
type::AnyRef can hold a non-owning reference to an instance of the native type of any thrift type tag. For example:
  int32_t value = 2;
  AnyRef ref = AnyRef::create<i32_t>(value);

The ref can then be used to manipulate the reference:
  ref.type(); // == i32_t{}
  ref.empty(); // false, 2 != 0
  ref.clear(); // Sets value to the default, 0
  ref.empty(); // true, now that value has been cleared.
  value; // 0

This is an improved version of thrift/conformance/cpp2/AnyRef.h, differing in the following ways:
 - Based on the Thrift type systems, not std::any.
 - Captures type-erased versions of the normal Thrift ops, which we then use to safely support those ops for type erased values.
 - Keeps track of the thrift type as well as the c++ type.
 - Does not have an extra 'unset' state that requires extra flow control and state checking. Instead a default initialized or reset AnyRef, is a reference to the Thrift Type system's native notion of nothing, a.k.a `void` or `thrift::type::void_t`

This is a bare bones API, but will unblock work on the runtime type registry, and we will refine the API as we build that out (for example, by potentially splitting mutable and const references into different classes, if it proves useful).

For example usage, see the use of `thrift::conformance::any_ref` in the thrift/conformance/cpp2 folder.

I've switched away from folly::Poly to get a bit more control on exactly what info is captured for a thrift type. I will update AnyValue in a subsequent diff to match, and support capturing an AnyValue in an AnyRef, so the types can interoperate.

Reviewed By: vitaut

Differential Revision: D32843221

fbshipit-source-id: 0ed99cca359bf87a64a6102265950ebd833f965f