suppress -Wignored-optimization-argument in generated cpp2 metadata
Summary: Some of these generated functions can get so large that the compiler gives up optimizing them. That would be alright - these functions are so rarely called that their being optimized is not particularly important - except that the compiler also emits a warning about it and this warning may be treated as an error. Suppress this warning so that builds do not fail when these functions get sufficiently large.

Reviewed By: Mizuchi, luciang

Differential Revision: D33592075

fbshipit-source-id: 5dfcf1383ae8722e1cdd53b813e32e0f05b57ae2