Use generator in range method of TestStreamServiceMock
Summary:
Use generator instead of publisher in range. For most tests this should make no difference but there are some tests (thrift/lib/cpp2/transport/rocket/test:streaming_tests - StreamingTest.SetMaxRequests) that use range and setChunkBufferSize(0) to simulate an ongoing stream. With StreamPublisher all the CPU thread work is done immediately and the results buffered. With generator there is real backpressure so the CPU work doesn't complete until the client consumes the data.

In our future refactoring the definition of active requests will change to requests running on the CPU worker pools so this change means the tests that use this and a requests limit will continue to work.

Reviewed By: praihan

Differential Revision: D33478320

fbshipit-source-id: 78c7dc1a26e105ceb442cb5602710b8f31dcb98b