Introduce rocket protocol MIME types
Summary: This diff introduces rocket MIME types for payload and metadata and enables using them with the C++ Rocket client and server. The presence of these MIME types in the setup frame makes the use of the rocket protocol key unnecessary and so it can be made optional. This may slightly simplify rocket client implementations in other languages, but also make Rocket protocol use RSocket features properly.

Reviewed By: stevegury, iahs

Differential Revision: D33595102

fbshipit-source-id: f85831918bda24e2723a68d0c3bb881dd605b226