Do not execute sslcontextobserver callback in forked child
Summary:
This callback uses folly::EventBase::runInEventBaseThread, which does not work
when it is invoked from a different pid than when the EventBase itself was
created.

Because this callback is triggered asynchronously by some arbitrary Observer
(which is backed by the Thrift flags module), it is possible for this code to
be invoked in the context of a forked child. Executing the callback without this
code would lead to the child aborting.

Reviewed By: andriigrynenko, AjanthanAsogamoorthy

Differential Revision: D33718481

fbshipit-source-id: deb7bdd75fec548469ec63e80c8a8a6c0fbb61ce