(standard) RoundRobinRequestPile implementation
Summary:
RoundRobinRequestPile will serve as the standard (default) RequestPile for most of the services.
It controls the request ordering by two dimensions - **priority and bucket**

Users are allowed to push into the queue with specified priority and bucket, and when the request is pulled, the ordering follows the following rules:
1. Higher priority queue will be checked first
2. For each priority, the requests in different buckets will be pulled in a round robin manner.

In order to optimize cpu usage, instead of using a bookmarking method for round robin recording, we used a composition queue method, in which each pull operation on a bucket will schedule another pull operation to the end of the bucket queue, if there are remaining requests.

Reviewed By: roddym

Differential Revision: D30554108

fbshipit-source-id: b27358a4f4234769e7daddf0b315c5d36041117d