Improve credit estimation in toAsyncGeneratorWithSizeTarget()
Summary:
The current algorithm for credit estimation uses the maximum payload size seen so far to guess how many future payloads can fit in the buffer.

This means that a single outlier payload can permanently degrade the throughput by consistently underestimating the number of credits. On long-running streams this is very likely to happen.

This diff uses instead a circular window to estimate the recent average payload size, which for any reasonable distribution of sizes still tracks the target buffer size but it is robust to outliers. Until enough payloads have been seen, the max is still conservatively used.

Reviewed By: iahs

Differential Revision: D33884510

fbshipit-source-id: 05ad2c7004e888cb9d989415e1d448a790e30418