ThriftServer Allow using None for handler
Summary:
With the new Status/Monitoring/Control interface it would be helpful to be able to standup a thrift server using an EmptyAsyncProcessorFactory.

Instead of requiring that we construct a wrapper class around EmptyAsyncProcessorFactory and all the hassle involved with cython and that can of worms, if None is used as a Handler we set it to EmptyAsyncProcessorFactory.

While we are here, since this thrift server doesn't talk to python at all there is no reason to set its thread manager to be the python executor if we are using an empty processor.

Reviewed By: pyamane

Differential Revision: D33858704

fbshipit-source-id: acfc35a7d0baee4d40237ab2b1a6245a5815149d