Update the default value for queue timeout flag to 100 ms
Summary:
We'd like to retire the `server_default_queue_timeout_ms` flag since it
was only used for rolling out the new 100 ms default and set the new default in
code directly. The queue timeout attribute will still remain dynamic but won't
be backed by this flag any longer. Instead, we should rely on `ThriftServerConfig`
as the source of truth, which can be set by each service independently. The default
will be 100 ms if services don't update that value in the config. We can even consider
disallowing this value from being set to 0 in the future.

Reviewed By: praihan

Differential Revision: D33821244

fbshipit-source-id: 562a66fa0d483f0ef39e5692af023fc35fb06925