fix c+11 narrowing warning in container_size test when running on Windows
Summary:
This was revealed in D33945540, V1. On Windows the data type for `off_t` is only 32-bit, so this leads to wraparound. Likely this works correctly if the type gets cast back to unsigned, but strictly speaking that's UB.

`off_t` is used by `folly::MemoryMapping` in case someone wants to change the API there to use 64-bit.

To fix, I'm opting to use a smaller memory size when running on Windows. 1 GiB should also be large enough

Reviewed By: yfeldblum

Differential Revision: D33948886

fbshipit-source-id: 91f27b4937c8903eae0025c72caa096884547f58