Fix race condition in ScopedServerThread
Summary:
ScopedServerThread sets the ThriftServer's event handler in `preServe()` which runs on the server thread, meanwhile `ThriftRocketServerHandler` constructor can access the list of the event handlers on a separate thread.

This diff removes the event handler manipulation from `preServe()`. Instead, the `ScopedServerThread` event handler is added to the list of server event handlers. As a consequence, two calls to the deprecated `setServerEventHandler()` function are removed.

Reviewed By: praihan

Differential Revision: D33951089

fbshipit-source-id: a2cdcc0b34bae5cb63b45137396201a3360b51a3