Fix shift-reduce errors in Thrift function definition
Summary:
This resolves shift-reduce conflicts related to the State 162 from P478404936.

The ambuguity boils down to the resolution of FunctionQualifier FunctionType part of the Function rule, when one of the tok_oneway, tok_idempotent or tok_readonly is encountered:
1) Resolve the token as FunctionQualifier
2) Resolve "" as FunctionQualifier and resolve the token as FunctionType.

The option 2 is allowed via rules FunctionType -> FieldType -> Identifier. Identifier makes exceptions for certain tokens for backwards compatibility.

The fix is to restrict exceptions to the identifier which can be used as a field/param/function type. Valid types with names of keywords can theoretically appear via a struct definition, e.g. "struct server {...}". When looking with zbgs, there are some usages of these but none has lowercase (e.g. these "struct Server {...}"). Thrift is case sensitive for both keywords and identifiers, so no exceptions has to be made for types.

Note: Yacc state report can be obtained by adding corresponding options to the TARGETS file D21797696.

ignore-unit-failures

Reviewed By: Alfus

Differential Revision: D34021808

fbshipit-source-id: baf1a5982526315cb0bebaeea8766f565dea9309