Simplify threshold calculation for credits request in toAsyncGeneratorWithSizeTarget()
Summary:
The calculation of `threshold` depends on the estimated payload size, but it is not necessary distinguish cases, the goal should be to send credit requests as frequently as possible, but at most every 16kB of incoming traffic, to amortize them.
We just need to handle small buffers, in which case we cannot rely on the amortization argument, so we send a request every half buffer.

Reviewed By: iahs

Differential Revision: D34083421

fbshipit-source-id: ceedfdab35abe22b6d6b96eff70b3812951df861