skip retrieval queue query for one bucket case
Summary:
There is one issue with synchronization between the RequestQueue and RetrievalQueue, which will lead us to empty dequeue() and leads to some expensive operations in the downstream ConcurrencyController.

This happens mostly when there is only one bucket per priority, so we are skipping the RetrievalQueue query for one bucket case (which is also an optimization as we don't need RetrievalQueue at all).

Reviewed By: roddym

Differential Revision: D33841285

fbshipit-source-id: 1a5801daf179d5584aa42ccb78cd991e3da4267f