Better incapsulation for attributes and annotations + shift-reduce conflicts fix
Summary:
Create separate annotated rules for enum values, definitions, fields and functions.

Separate Function into two rules: Function (only for regular functions) and Performs (for performs clauses).

List rules of aforementioned items are simplified to use the annotated rules and separators.

A comma or a semicolon is now mandatory after performs clauses. Use cases already have it, so there's no need for a migration. This solves shift-reduce conflicts due to an ambiguity of annotations resolution - they might belong to either the field type from the performs clause or to the function definition (as performs clause is a type of a function here). I'm not sure annotations for performs clauses were even intended, but this was allowed due to how the rules were written. The refactoring above avoids similar problems in the future and makes the fix fit more nicely with the other rules structure.

Reviewed By: Alfus

Differential Revision: D34054862

fbshipit-source-id: 87edb8312d2aa3aa294512bd93a337b81a4d9692